No boost dependency anymore
Newer GT4Py versions ship with a vendored in version of the boost header
file that it needs. And since NDSL updated its GT4Py version, we don't
need to install boost anymore in the `Dockerfile` for PyFV3. Not sure if
anyone uses the docker version, but it's easy enough to just clean up.