ci: no test extra for NDSL in external trigger
Translate tests of pyFV3 run as hook in other repositories like NDSL.
When pyFV3 tests run as part of NDSL's CI, we use the NDSL version that
is about to be merged, to check that the changes in NDSL don't break
pyFV3. When we do this, we don't need to install NDSL with the "test"
extra because that should only be necessary to run NDSL tests (which we
don't run as part of pyFV3 CI).